add commands.hpp and graph impl

+++ sogdanov.sanal/S3/Graph.hpp
@@ -11,23 +11,20 @@ namespace sogdanov
   class Graph
   {
     Vector<std::string> vertices_;
-    HashTable<std::pair<std::string, std::string>, Vector<unsigned int>> edges_;
-
+    HashTable<std::pair<std::string, std::string>, Vector<size_t>> edges_;
     void internal_add_vertex(const std::string &v);
-
   public:
     Graph();
-
     void swap(Graph &other) noexcept;
 
-    void bind(const std::string &u, const std::string &v, unsigned int w);
-    void cut(const std::string &u, const std::string &v, unsigned int w);
+    void bind(const std::string &u, const std::string &v, size_t w);
+    void cut(const std::string &u, const std::string &v, size_t w);
     void add_vertex(const std::string &v);
 
     bool has_vertex(const std::string &v) const;
     Vector<std::string> get_vertices() const;
 
-    HashTable<std::pair<std::string, std::string>, Vector<unsigned int>> &get_edges();
+    HashTable<std::pair<std::string, std::string>, Vector<size_t>> &get_edges();
   };
 
 }
